+++ lib/astro/solar/sun_rise_set.ex
@@ -202,6 +202,10 @@ defmodule Astro.Solar.SunRiseSet do
   * `{:error, :time_zone_not_resolved}` if the time zone cannot be
     resolved from the location.
 
+  * `{:error, :tz_world_data_not_installed}` if `:tz_world` is a
+    dependency but its time zone data has not been downloaded. Run
+    `mix tz_world.update` to install it.
+
   ### Examples
 
       iex> moment = Astro.Time.date_time_to_moment(~D[2019-12-04])
@@ -260,6 +264,10 @@ defmodule Astro.Solar.SunRiseSet do
 
   * `{:error, :time_zone_not_resolved}` if the time zone cannot be
     resolved from the location.
+
+  * `{:error, :tz_world_data_not_installed}` if `:tz_world` is a
+    dependency but its time zone data has not been downloaded. Run
+    `mix tz_world.update` to install it.
 
   ### Examples
 
@@ -474,11 +482,7 @@ defmodule Astro.Solar.SunRiseSet do
     resolver.(%Geo.Point{coordinates: {lng, lat}})
   end
 
-  if Code.ensure_loaded?(TzWorld) do
-    defp default_resolver(point), do: TzWorld.timezone_at(point)
-  else
-    defp default_resolver(_point), do: {:error, :time_zone_not_resolved}
-  end
+  defp default_resolver(point), do: Time.tz_world_timezone_at(point)
 
   defp shift_zone(utc_dt, tz, :configured), do: DateTime.shift_zone(utc_dt, tz)
   defp shift_zone(utc_dt, tz, tz_db), do: DateTime.shift_zone(utc_dt, tz, tz_db)